requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
openai==1.6.1
PyPDF2==3.0.1
tiktoken==0.5.2